Added conditional Menubar rendering based on viewport size. Added a mobile dropdown menu bar.
